src/requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.5
dotenv==0.9.9
idna==3.11
kafka-python==2.3.0
psycopg2==2.9.11
python-dotenv==1.2.2
requests==2.32.5
urllib3==2.6.3
